Refactored unused code
